Frog speed increases with each reset
I've created a function to start the frog timer. The speed is a variable that gets changed every time the fly reset takes place. When it does, it changes he speed, stops the current timer, and starts the new one (with the updated speed). I am doing this with setInterval(moveFrog, frogSpeed).